Uncap watershed detail in the Node binding + sync docs
`6f163fc` widened watershed_detail from u8 to u32 in the core, CLI, and
Python binding but not the Node/wasm binding, which still assigned a u8 to
config.watershed_detail — a wasm build error (E0308). Widen the Node
Options field to u32 and refresh every "0..=255" watershed doc (CLI help,
Config field, .pyi, index.d.ts, and all three READMEs) to reflect that the
level is now uncapped (default 128; higher = more regions).

+++ nodejs/src/lib.rs
@@ -40,8 +40,9 @@ struct Options {
     adaptive_window: Option<u32>,
     /// Adaptive sensitivity: percent below the local mean (default 15).
     adaptive_t: Option<f64>,
-    /// Watershed clustering: hierarchy cut level (0..=255).
-    watershed_detail: Option<u8>,
+    /// Watershed clustering: hierarchy cut level (default 128; higher = more
+    /// regions, uncapped).
+    watershed_detail: Option<u32>,
     /// One of "bw" | "poster" | "photo"; applied before the other fields.
     preset: Option<String>,
 }
